Add files via upload
this project aims to investigate a data set with the data analysis process as an introduction to data analysis . The project give me 5 options of data sets to work at , I selected No-show appointments data set , which is collects information from 100-k medical appointments from Brazil and is focused on the question of whether or not patients show up for their appointment. you can find data set options in data-set-options.pdf file.